+++ database/setup.py
@@ -5,7 +5,7 @@ package_name = 'database'
 setup(
     name=package_name,
     version='0.0.0',
-    packages=[package_name],
+    packages=[package_name, 'protocol_handler'],
     data_files=[
         ('share/ament_index/resource_index/packages',
             ['resource/' + package_name]),
@@ -15,14 +15,15 @@ setup(
     zip_safe=True,
     maintainer='Doga',
     maintainer_email='dozgulbas@anl.gov',
-    description='Database files for the database',
+    description='Database and protocol parser files for communication between ROS and database and ROS and internal pis',
     license='MIT Licenses',
     tests_require=['pytest'],
     entry_points={
         'console_scripts': [
             'database_functions = database.database_functions:main_null',
             'connect = database.connect:main_null',
-            'protocol_parser = database.protocol_parser:main_null'
+            'protocol_parser = protocol_handler.protocol_parser:main_null',
+            'protocol_handling_client = protocol_handler.protocol_handling_client:main_null'
         ],
     },
 )
